Dados em tempo real melhorado + dados de humidade

+++ src/main/java/tfc/plataforma/projeto_v1/BuildingData.java
@@ -3,11 +3,11 @@ package tfc.plataforma.projeto_v1;
 public class BuildingData {
     private String id;
     private String data;
-    private float temp;
-    private float luminosidade;
-    private float humidade;
+    private double temp;
+    private double luminosidade;
+    private double humidade;
 
-    public BuildingData(String id, String data, float temp, float luminosidade, float humidade){
+    public BuildingData(String id, String data, double temp, double luminosidade, double humidade){
         this.id = id;
         this.data = data;
         this.temp = temp;
